+++ Controllers/TableController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 using RestaurangWebAPI.Models;
+using System.Net.Http.Headers;
 
 namespace RestaurangWebAPI.Controllers
 {
@@ -54,6 +55,15 @@ namespace RestaurangWebAPI.Controllers
         {
             Console.WriteLine(model.Seats);
 
+            var token = Request.Cookies["jwtToken"];
+            Console.WriteLine(token);
+            if (token == null)
+            {
+                return RedirectToPage("/Login");
+            }
+
+            _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+
             if (ModelState.IsValid)
             {
 
@@ -100,6 +110,15 @@ namespace RestaurangWebAPI.Controllers
                 return View(model);
             }
 
+            var token = Request.Cookies["jwtToken"];
+            Console.WriteLine(token);
+            if (token == null)
+            {
+                return RedirectToPage("/Login");
+            }
+
+            _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+
             var response = await _httpClient.PatchAsJsonAsync(_httpClient.BaseAddress + "/Table/UpdateTable", model);
             if (response.IsSuccessStatusCode)
             {
@@ -131,6 +150,15 @@ namespace RestaurangWebAPI.Controllers
         [HttpPost, ActionName("Delete")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            var token = Request.Cookies["jwtToken"];
+            Console.WriteLine(token);
+            if (token == null)
+            {
+                return RedirectToPage("/Login");
+            }
+
+            _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+
             var response = await _httpClient.DeleteAsync(_httpClient.BaseAddress + $"/Table/{id}");
             if (response.IsSuccessStatusCode)
             {
